* master: (666 commits)
  add okta support
  Merge branch 'master' of github.com:apereo/cas
  clean up config to handle capacity fields
  Add annotation processing for logging so that log4j file is created (#4516)
  working on test cases for mfa webflow
  working on test cases for mfa webflow
  update jradius dependency to use version from maven central (#4507)
  add support for azuread authn
  renovatebot(deps): update spring data to v2.2.3.release
  fix error-prone
  remove groovy check
  Ehcache 3 ticket registry and Ehcache 3 used in X509 CRL cache (#4488)
  renovatebot(deps): update micrometer to v1.3.2
  renovatebot(deps): update dependency org.webjars:bootstrap to v4.4.1 (#4509)
  Jpa: Simple MFA configuration should change TransientSessionTicket to TransientSessionTicketImpl (#4502)
  Limit stack traces (#4503)
  clean up codebase with new error-prone
  add tests for trusted webflow mfa config
  add tests for trusted webflow mfa config
  add tests for trusted webflow mfa config
  ...

# Conflicts:
#	core/cas-server-core-services-registry/src/test/java/org/apereo/cas/services/resource/ModifyResourceBasedRegisteredServiceWatcherTests.java
#	support/cas-server-support-acceptto-mfa/src/main/java/org/apereo/cas/config/AccepttoMultifactorAuthenticationConfiguration.java

+++ support/cas-server-support-redis-ticket-registry/src/main/java/org/apereo/cas/ticket/registry/RedisTicketRegistry.java
@@ -88,7 +88,7 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
             val redisKey = getTicketRedisKey(encodeTicketId(ticket.getId()));
             val encodeTicket = encodeTicket(ticket);
             val timeout = getTimeout(ticket);
-            this.client.boundValueOps(redisKey).set(encodeTicket, timeout.longValue(), TimeUnit.SECONDS);
+            this.client.boundValueOps(redisKey).set(encodeTicket, timeout, TimeUnit.SECONDS);
         } catch (final Exception e) {
             LOGGER.error("Failed to add [{}]", ticket, e);
         }
@@ -146,7 +146,7 @@ public class RedisTicketRegistry extends AbstractTicketRegistry {
             LOGGER.debug("Fetched redis key [{}] for ticket [{}]", redisKey, ticket);
 
             val timeout = getTimeout(ticket);
-            this.client.boundValueOps(redisKey).set(encodeTicket, timeout.longValue(), TimeUnit.SECONDS);
+            this.client.boundValueOps(redisKey).set(encodeTicket, timeout, TimeUnit.SECONDS);
             return encodeTicket;
         } catch (final Exception e) {
             LOGGER.error("Failed to update [{}]", ticket, e);
